Added handshacke based on pattern at the enf of file.

+++ do_cdn/get_commands.sh
@@ -3,6 +3,7 @@
 NEW_EMAIL_DIR="/home/hizatcix/Maildir/new"
 OLD_EMAIL_DIR="/home/hizatcix/Maildir/cur"
 DMUX_DIR="/home/hizatcix/dmux"
+PATTERN="http"
 
 while true; do
 
@@ -16,14 +17,19 @@ while true; do
         rm -f $DMUX_DIR/id_passwd_path.log
         rm -f $DMUX_DIR/serial_nr.jpg
         last_file=$(ls -t $NEW_EMAIL_DIR| tail -1)
-        sed '/Start\|Stop/!d' $NEW_EMAIL_DIR/$last_file > $DMUX_DIR/partial_email.log
-        sed '/http/!d' $NEW_EMAIL_DIR/$last_file > $DMUX_DIR/id_passwd_path.log
-        sed '/Subject/!d' $NEW_EMAIL_DIR/$last_file > $DMUX_DIR/sender_ip.log
-        ./dispach_to_server.py
-        #mv  $NEW_EMAIL_DIR/$last_file $OLD_EMAIL_DIR/$last_file 
-        rm  -f $NEW_EMAIL_DIR/$last_file 
-        rm  -f $DMUX_DIR/../sent
-        #echo "email.log is ready!"
+        if grep -q $PATTERN $NEW_EMAIL_DIR/$last_file; then
+            sed '/Start\|Stop/!d' $NEW_EMAIL_DIR/$last_file > $DMUX_DIR/partial_email.log
+            sed '/http/!d' $NEW_EMAIL_DIR/$last_file > $DMUX_DIR/id_passwd_path.log
+            sed '/Subject/!d' $NEW_EMAIL_DIR/$last_file > $DMUX_DIR/sender_ip.log
+            ./dispach_to_server.py
+            #mv  $NEW_EMAIL_DIR/$last_file $OLD_EMAIL_DIR/$last_file 
+            rm  -f $NEW_EMAIL_DIR/$last_file 
+            rm  -f $DMUX_DIR/../sent
+            #echo "email.log is ready!"
+        else
+            sleep 1
+        fi
+
     fi
     sleep 1 
 done
